lab3/requirements.txt
alembic==1.15.2
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.9
MarkupSafe==3.0.2
marshmallow==3.26.1
packaging==24.2
psycopg2-binary==2.9.10
python-dotenv==1.1.0
SQLAlchemy==2.0.39
typing_extensions==4.13.0
Werkzeug==3.1.3
zipp==3.21.0
